Update Alter and add AlterExpression class for multiple ADD/DROP expressions in a single ALTER statement
Update .jjt file to break out the AlterExpression into its own class and for Alter to compose multiple AlterExpressions

Update AlterTest for new AlterExpressions and test for multiple ADD/DROP statements in a single ALTER
