+++ src/main/java/net/sf/jsqlparser/statement/alter/Alter.java
@@ -22,7 +22,8 @@
 package net.sf.jsqlparser.statement.alter;
 
 import java.util.ArrayList;
-import java.util.List;
+import java.util.Iterator;
+
 import net.sf.jsqlparser.schema.Table;
 import net.sf.jsqlparser.statement.Statement;
 import net.sf.jsqlparser.statement.StatementVisitor;
@@ -32,95 +33,13 @@ import net.sf.jsqlparser.statement.select.PlainSelect;
 
 /**
  *
- * @author toben
+ * @author toben & wrobstory
  */
 public class Alter implements Statement {
 
     private Table table;
-    private String columnName;
-    //private ColDataType dataType;
 
-    private List<ColumnDataType> colDataTypeList;
-
-    private List<String> pkColumns;
-    private List<String> ukColumns;
-    private String ukName;
-    private Index index = null;
-    private AlterOperation operation;
-    private String constraintName;
-    private boolean onDeleteRestrict;
-    private boolean onDeleteSetNull;
-    private boolean onDeleteCascade;
-    private List<String> fkColumns;
-    private String fkSourceTable;
-    private List<String> fkSourceColumns;
-
-    public boolean isOnDeleteCascade() {
-        return onDeleteCascade;
-    }
-
-    public void setOnDeleteCascade(boolean onDeleteCascade) {
-        this.onDeleteCascade = onDeleteCascade;
-    }
-
-    public boolean isOnDeleteRestrict() {
-        return onDeleteRestrict;
-    }
-
-    public void setOnDeleteRestrict(boolean onDeleteRestrict) {
-        this.onDeleteRestrict = onDeleteRestrict;
-    }
-
-    public boolean isOnDeleteSetNull() {
-        return onDeleteSetNull;
-    }
-
-    public void setOnDeleteSetNull(boolean onDeleteSetNull) {
-        this.onDeleteSetNull = onDeleteSetNull;
-    }
-
-    public List<String> getFkColumns() {
-        return fkColumns;
-    }
-
-    public void setFkColumns(List<String> fkColumns) {
-        this.fkColumns = fkColumns;
-    }
-
-    public String getFkSourceTable() {
-        return fkSourceTable;
-    }
-
-    public void setFkSourceTable(String fkSourceTable) {
-        this.fkSourceTable = fkSourceTable;
-    }
-
-    public List<ColumnDataType> getColDataTypeList() {
-        return colDataTypeList;
-    }
-
-    public void addColDataType(String columnName, ColDataType colDataType) {
-        if (colDataTypeList == null) {
-            colDataTypeList = new ArrayList<ColumnDataType>();
-        }
-        colDataTypeList.add(new ColumnDataType(columnName, colDataType));
-    }
-
-    public AlterOperation getOperation() {
-        return operation;
-    }
-
-    public void setOperation(AlterOperation operation) {
-        this.operation = operation;
-    }
-
-    public List<String> getFkSourceColumns() {
-        return fkSourceColumns;
-    }
-
-    public void setFkSourceColumns(List<String> fkSourceColumns) {
-        this.fkSourceColumns = fkSourceColumns;
-    }
+    private ArrayList<AlterExpression> alterExpressions;
 
     public Table getTable() {
         return table;
@@ -130,44 +49,19 @@ public class Alter implements Statement {
         this.table = table;
     }
 
-    public String getColumnName() {
-        return columnName;
+    public void addAlterExpression(AlterExpression alterExpression){
+        if (alterExpressions == null) {
+            alterExpressions = new ArrayList<AlterExpression>();
+        }
+        alterExpressions.add(alterExpression);
     }
 
-    public void setColumnName(String columnName) {
-        this.columnName = columnName;
+    public ArrayList<AlterExpression> getAlterExpressions() {
+        return alterExpressions;
     }
 
-    public String getConstraintName() {
-        return this.constraintName;
-    }
-
-    public void setConstraintName(final String constraintName) {
-        this.constraintName = constraintName;
-    }
-
-    public List<String> getPkColumns() {
-        return pkColumns;
-    }
-
-    public void setPkColumns(List<String> pkColumns) {
-        this.pkColumns = pkColumns;
-    }
-
-    public List<String> getUkColumns() {
-        return ukColumns;
-    }
-
-    public void setUkColumns(List<String> ukColumns) {
-        this.ukColumns = ukColumns;
-    }
-
-    public String getUkName() {
-        return ukName;
-    }
-
-    public void setUkName(String ukName) {
-        this.ukName = ukName;
+    public void setAlterExpressions(ArrayList<AlterExpression> alterExpressions) {
+        this.alterExpressions = alterExpressions;
     }
 
     @Override
@@ -175,71 +69,25 @@ public class Alter implements Statement {
         statementVisitor.visit(this);
     }
 
-    public Index getIndex() {
-        return index;
-    }
-
-    public void setIndex(Index index) {
-        this.index = index;
-    }
-
     @Override
     public String toString() {
+
         StringBuilder b = new StringBuilder();
-        b.append("ALTER TABLE ").append(table.getFullyQualifiedName()).append(" ").append(operation).append(" ");
-        if (columnName != null) {
-            b.append("COLUMN ").append(columnName);
-        } else if (colDataTypeList != null) {
-            if (colDataTypeList.size() > 1) {
-                b.append("(");
-            } else b.append("COLUMN ");
-            b.append(PlainSelect.getStringList(colDataTypeList));
-            if (colDataTypeList.size() > 1) {
-                b.append(")");
+        b.append("ALTER TABLE ").append(table.getFullyQualifiedName()).append(" ");
+
+        Iterator<AlterExpression> altIter = alterExpressions.iterator();
+
+        while (altIter.hasNext()){
+            b.append(altIter.next().toString());
+
+            // Need to append whitespace after each ADD or DROP statement
+            // but not the last one
+            if (altIter.hasNext()){
+                b.append(", ");
             }
-        } else if (constraintName != null) {
-            b.append("CONSTRAINT ").append(constraintName);
-        } else if (pkColumns != null) {
-            b.append("PRIMARY KEY (").append(PlainSelect.getStringList(pkColumns)).append(")");
-        } else if (ukColumns != null) {
-            b.append("UNIQUE KEY ").append(ukName).append(" (").append(PlainSelect.getStringList(ukColumns)).append(")");
-        } else if (fkColumns != null) {
-            b.append("FOREIGN KEY (").append(PlainSelect.getStringList(fkColumns)).append(") REFERENCES ").append(fkSourceTable).append(" (").append(
-                    PlainSelect.getStringList(fkSourceColumns)).append(")");
-            if (isOnDeleteCascade()) {
-                b.append(" ON DELETE CASCADE");
-            } else if (isOnDeleteRestrict()) {
-                b.append(" ON DELETE RESTRICT");
-            } else if (isOnDeleteSetNull()) {
-                b.append(" ON DELETE SET NULL");
-            }
-        } else if (index != null) {
-            b.append(index);
         }
+
         return b.toString();
     }
 
-    public class ColumnDataType {
-
-        private final String columnName;
-        private final ColDataType colDataType;
-
-        public ColumnDataType(String columnName, ColDataType colDataType) {
-            this.columnName = columnName;
-            this.colDataType = colDataType;
-        }
-
-        public String getColumnName() {
-            return columnName;
-        }
-
-        public ColDataType getColDataType() {
-            return colDataType;
-        }
-
-        @Override
-        public String toString() {
-            return columnName + " " + colDataType;
-        }
-    }
 }
